Hierarchy reorganization
* Added new strong name
* Moved all serialization and deserialization support to separate
namespace
