+++ Cyotek.Data.Nbt.Tests/XmlTagWriterTests.cs
@@ -1,4 +1,5 @@
 using System.IO;
+using Cyotek.Data.Nbt.Serialization;
 using NUnit.Framework;
 
 namespace Cyotek.Data.Nbt.Tests
