+++ Cyotek.Data.Nbt/NbtDocument.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using System.IO.Compression;
 using System.Text;
+using Cyotek.Data.Nbt.Serialization;
 
 namespace Cyotek.Data.Nbt
 {
