+++ Cyotek.Data.Nbt/Properties/AssemblyInfo.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Cyotek NBT")]
@@ -14,6 +13,5 @@ using System.Runtime.InteropServices;
 [assembly: CLSCompliant(true)]
 [assembly: ComVisible(false)]
 [assembly: Guid("0D7BAD0E-1765-4D49-A31F-430E76638B49")]
-[assembly: InternalsVisibleTo("Cyotek.Data.Nbt.Tests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100f711fe95c698a0533975dc0eabc40db1876c95e7a4b03ef65765b7f10d81b71c0f64d4d2ae7c1026d6426fa2d09c2dc16766acfa397141c1496dad3ee8be06749fb655edd7e4c902ad5a87074f56c0a452bc181a5fa8e3e4a441ecd747921e13d239737f4fdffdeabeb09e5386bbb865d41b560b787dbc9b9d29e3aa0b4b1ac4")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.4.5")]
+[assembly: AssemblyVersion("2.0.0.0")]
+[assembly: AssemblyFileVersion("2.0.0.0")]
