+++ Cyotek.Data.Nbt/Serialization/ITagWriter.cs
@@ -1,6 +1,6 @@
 using System.IO;
 
-namespace Cyotek.Data.Nbt
+namespace Cyotek.Data.Nbt.Serialization
 {
   public interface ITagWriter
   {
