+++ Cyotek.Data.Nbt/Serialization/XmlTagWriter.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Text;
 using System.Xml;
 
-namespace Cyotek.Data.Nbt
+namespace Cyotek.Data.Nbt.Serialization
 {
   public class XmlTagWriter : ITagWriter
   {
