+++ Cyotek.Data.Nbt/Tag.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.IO;
 using System.Text;
+using Cyotek.Data.Nbt.Serialization;
 
 namespace Cyotek.Data.Nbt
 {
